refactor app for better structure
