+++ src/client/src/main.ts
@@ -7,7 +7,7 @@ import { bootstrap } from '@angular/platform-browser-dynamic';
 import { enableProdMode } from '@angular/core';
 import { HTTP_PROVIDERS } from '@angular/http';
 import { provideRouter, RouterConfig } from '@angular/router';
-import { AppComponent } from './app.component';
+import { AppComponent } from './components/app.component';
 import { appRoutes } from './routes/main.routes';
 import { ConfirmResolver } from './resolvers/confirm.resolver';
 import { EncryptionService } from './services/encryption.service';
